refactor: improves code organization
