+++ helios/__init__.py
@@ -1,0 +1,3 @@
+from helios.backends.fury.draw import NetworkDraw
+
+__all__ = ['NetworkDraw']
