--- helios/fury/__init__.py
@@ -1,3 +1,0 @@
-from helios.fury.actors import NetworkDraw
-
-__all__ = ['NetworkDraw']
