+++ helios/layouts/__init__.py
@@ -1,0 +1,3 @@
+from helios.layouts.force_directed import HeliosFr
+
+__all__ = ['HeliosFr']
